Rename mean_lp_error.rs to lp_error.rs
