+++ burn-book/src/building-blocks/module.md
@@ -295,6 +295,6 @@ Burn comes with built-in modules that you can use to build your own modules.
 | `CosineEmbeddingLoss`    | `nn.CosineEmbeddingLoss` |
 | `CrossEntropyLoss`       | `nn.CrossEntropyLoss`    |
 | `HuberLoss`              | `nn.HuberLoss`           |
-| `MeanLpErrorLoss`        | _No direct equivalent_   |
+| `LpLoss`                 | _No direct equivalent_   |
 | `MseLoss`                | `nn.MSELoss`             |
 | `PoissonNllLoss`         | `nn.PoissonNLLLoss`      |
