+++ crates/burn-nn/src/loss/lp_loss.rs
