+++ crates/burn-nn/src/loss/mod.rs
@@ -2,7 +2,7 @@ mod binary_cross_entropy;
 mod cosine_embedding;
 mod cross_entropy;
 mod huber;
-mod mean_lp_error;
+mod lp_loss;
 mod mse;
 mod poisson;
 mod reduction;
@@ -11,7 +11,7 @@ pub use binary_cross_entropy::*;
 pub use cosine_embedding::*;
 pub use cross_entropy::*;
 pub use huber::*;
-pub use mean_lp_error::*;
+pub use lp_loss::*;
 pub use mse::*;
 pub use poisson::*;
 pub use reduction::*;
